refactor: Extract interface for FirebaseHandler
